+++ cmd/forklift/dev/plt/cli.go
@@ -1,14 +1,15 @@
-// Package env provides subcommands for the development environment
-package env
+// Package plt provides subcommands for the development pallet
+package plt
 
 import (
 	"github.com/urfave/cli/v2"
 )
 
 var Cmd = &cli.Command{
-	Name:    "env",
-	Aliases: []string{"environment"},
-	Usage:   "Facilitates development and maintenance of a Forklift environment",
+	Name:    "plt",
+	Aliases: []string{"pallet"},
+	Usage: "Facilitates development and maintenance of a Forklift pallet in the current working " +
+		"directory",
 	Flags: []cli.Flag{
 		&cli.StringSliceFlag{
 			Name: "repos",
@@ -20,93 +21,92 @@ var Cmd = &cli.Command{
 		{
 			Name:     "cache-repo",
 			Aliases:  []string{"cache-repositories"},
-			Category: "Use the environment",
-			Usage:    "Updates the cache with the repos available in the development environment",
+			Category: "Use the pallet",
+			Usage:    "Updates the cache with the repos available in the development pallet",
 			Action:   cacheRepoAction,
 		},
 		{
 			Name:     "cache-img",
 			Aliases:  []string{"cache-images"},
-			Category: "Use the environment",
-			Usage:    "Pre-downloads the Docker container images required by the development environment",
+			Category: "Use the pallet",
+			Usage:    "Pre-downloads the Docker container images required by the development pallet",
 			Action:   cacheImgAction,
 		},
 		{
 			Name:     "check",
-			Category: "Use the environment",
-			Usage:    "Checks whether the development environment's resource constraints are satisfied",
+			Category: "Use the pallet",
+			Usage:    "Checks whether the development pallet's resource constraints are satisfied",
 			Action:   checkAction,
 		},
 		{
 			Name:     "plan",
-			Category: "Use the environment",
-			Usage: "Determines the changes needed to update the Docker Swarm to match the deployments " +
-				"specified by the local environment",
+			Category: "Use the pallet",
+			Usage: "Determines the changes needed to update the Docker host to match the deployments " +
+				"specified by the local pallet",
 			Action: planAction,
 		},
 		{
 			Name:     "apply",
-			Category: "Use the environment",
-			Usage: "Updates the Docker Swarm to match the deployments specified by the " +
-				"development environment",
-			Action: applyAction,
+			Category: "Use the pallet",
+			Usage:    "Updates the Docker host to match the deployments specified by the development pallet",
+			Action:   applyAction,
 		},
 		{
 			Name:     "show",
-			Category: "Query the environment",
-			Usage:    "Describes the development environment",
+			Category: "Query the pallet",
+			Usage:    "Describes the development pallet",
 			Action:   showAction,
 		},
 		{
 			Name:     "ls-repo",
 			Aliases:  []string{"list-repositories"},
-			Category: "Query the environment",
-			Usage:    "Lists repos specified by the environment",
+			Category: "Query the pallet",
+			Usage:    "Lists repos specified by the pallet",
 			Action:   lsRepoAction,
 		},
 		{
 			Name:      "show-repo",
 			Aliases:   []string{"show-repository"},
-			Category:  "Query the environment",
-			Usage:     "Describes a repo available in the development environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a repo available in the development pallet",
 			ArgsUsage: "repo_path",
 			Action:    showRepoAction,
 		},
 		{
 			Name:     "ls-pkg",
 			Aliases:  []string{"list-packages"},
-			Category: "Query the environment",
-			Usage:    "Lists packages available in the development environment",
+			Category: "Query the pallet",
+			Usage:    "Lists packages available in the development pallet",
 			Action:   lsPkgAction,
 		},
 		{
 			Name:      "show-pkg",
 			Aliases:   []string{"show-package"},
-			Category:  "Query the environment",
-			Usage:     "Describes a package available in the development environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a package available in the development pallet",
 			ArgsUsage: "package_path",
 			Action:    showPkgAction,
 		},
 		{
 			Name:     "ls-depl",
 			Aliases:  []string{"list-deployments"},
-			Category: "Query the environment",
-			Usage:    "Lists package deployments specified by the development environment",
+			Category: "Query the pallet",
+			Usage:    "Lists package deployments specified by the development pallet",
 			Action:   lsDeplAction,
 		},
 		{
 			Name:      "show-depl",
 			Aliases:   []string{"show-deployment"},
-			Category:  "Query the environment",
-			Usage:     "Describes a package deployment specified by the development environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a package deployment specified by the development pallet",
 			ArgsUsage: "package_path",
 			Action:    showDeplAction,
 		},
 		{
 			Name:      "add-repo",
 			Aliases:   []string{"add-repositories"},
-			Category:  "Modify the environment",
-			Usage:     "Adds repos to the environment, tracking specified versions or branches",
+			Category:  "Modify the pallet",
+			Usage:     "Adds repos to the pallet, tracking specified versions or branches",
 			ArgsUsage: "[repo_path@version_query]...",
 			Action:    addRepoAction,
 		},
@@ -114,8 +114,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:      "rm-repo",
 		// 	Aliases:   []string{"remove-repositories},
-		// 	Category:  "Modify the environment",
-		// 	Usage:     "Removes a repo from the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage:     "Removes a repo from the pallet",
 		// 	ArgsUsage: "repo_path",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("removing repo", c.Args().First())
@@ -125,8 +125,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:  "add-depl",
 		// 	Aliases:   []string{"add-deployments"},
-		// 	Category:  "Modify the environment",
-		// 	Usage: "Adds a package deployment to the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage: "Adds a package deployment to the pallet",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("adding package deployment", c.Args().First())
 		// 		return nil
@@ -135,8 +135,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:      "rm-depl",
 		// 	Aliases:   []string{"remove-deployments"},
-		// 	Category:  "Modify the environment",
-		// 	Usage:     "Removes a package deployment from the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage:     "Removes a package deployment from the pallet",
 		// 	ArgsUsage: "package_path",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("removing package deployment", c.Args().First())