+++ cmd/forklift/host/host.go
@@ -1,4 +1,4 @@
-package depl
+package host
 
 import (
 	"context"