+++ cmd/forklift/main.go
@@ -8,9 +8,9 @@ import (
 	"github.com/urfave/cli/v2"
 
 	"github.com/PlanktoScope/forklift/cmd/forklift/cache"
-	"github.com/PlanktoScope/forklift/cmd/forklift/depl"
 	"github.com/PlanktoScope/forklift/cmd/forklift/dev"
-	"github.com/PlanktoScope/forklift/cmd/forklift/env"
+	"github.com/PlanktoScope/forklift/cmd/forklift/host"
+	"github.com/PlanktoScope/forklift/cmd/forklift/plt"
 )
 
 func main() {
@@ -28,9 +28,9 @@ var app = &cli.App{
 	Version: "v0.2.2",
 	Usage:   "Manages pallets and package deployments",
 	Commands: []*cli.Command{
-		env.Cmd,
+		plt.Cmd,
 		cache.Cmd,
-		depl.Cmd,
+		host.Cmd,
 		dev.Cmd,
 	},
 	Flags: []cli.Flag{