+++ cmd/forklift/plt/cli.go
@@ -1,44 +1,44 @@
-// Package env provides subcommands for the local environment
-package env
+// Package plt provides subcommands for the local pallet
+package plt
 
 import (
 	"github.com/urfave/cli/v2"
 )
 
 var Cmd = &cli.Command{
-	Name:    "env",
-	Aliases: []string{"environment"},
-	Usage:   "Manages the local environment",
+	Name:    "plt",
+	Aliases: []string{"pallet"},
+	Usage:   "Manages the local pallet",
 	Subcommands: []*cli.Command{
 		{
 			Name:      "clone",
-			Category:  "Modify the environment",
-			Usage:     "Initializes the local environment from a remote release",
+			Category:  "Modify the pallet",
+			Usage:     "Initializes the local pallet from a remote release",
 			ArgsUsage: "[github_repository_path@release]",
 			Flags: []cli.Flag{
 				&cli.BoolFlag{
 					Name:  "force",
-					Usage: "Deletes the local environment if it already exists",
+					Usage: "Deletes the local pallet if it already exists",
 				},
 			},
 			Action: cloneAction,
 		},
 		{
 			Name:     "fetch",
-			Category: "Modify the environment",
+			Category: "Modify the pallet",
 			Usage:    "Updates information about the remote release",
 			Action:   fetchAction,
 		},
 		{
 			Name:     "pull",
-			Category: "Modify the environment",
-			Usage:    "Fast-forwards the local environment to match the remote release",
+			Category: "Modify the pallet",
+			Usage:    "Fast-forwards the local pallet to match the remote release",
 			Action:   pullAction,
 		},
 		// {
 		// 	Name:  "push",
-		// 	Category:  "Modify the environment",
-		// 	Usage: "Updates the remote release from the local environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage: "Updates the remote release from the local pallet",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("pushing to remote origin")
 		// 		return nil
@@ -47,101 +47,100 @@ var Cmd = &cli.Command{
 		{
 			Name:     "rm",
 			Aliases:  []string{"remove"},
-			Category: "Modify the environment",
-			Usage:    "Removes the local environment",
+			Category: "Modify the pallet",
+			Usage:    "Removes the local pallet",
 			Action:   rmAction,
 		},
 		// remoteCmd,
 		{
 			Name:     "cache-repo",
 			Aliases:  []string{"cache-repositories"},
-			Category: "Use the environment",
-			Usage:    "Updates the cache with the repos available in the local environment",
+			Category: "Use the pallet",
+			Usage:    "Updates the cache with the repos available in the local pallet",
 			Action:   cacheRepoAction,
 		},
 		{
 			Name:     "cache-img",
 			Aliases:  []string{"cache-images"},
-			Category: "Use the environment",
-			Usage:    "Pre-downloads the Docker container images required by the local environment",
+			Category: "Use the pallet",
+			Usage:    "Pre-downloads the Docker container images required by the local pallet",
 			Action:   cacheImgAction,
 		},
 		{
 			Name:     "check",
-			Category: "Use the environment",
-			Usage:    "Checks whether the local environment's resource constraints are satisfied",
+			Category: "Use the pallet",
+			Usage:    "Checks whether the local pallet's resource constraints are satisfied",
 			Action:   checkAction,
 		},
 		{
 			Name:     "plan",
-			Category: "Use the environment",
-			Usage: "Determines the changes needed to update the Docker Swarm to match the deployments " +
-				"specified by the local environment",
+			Category: "Use the pallet",
+			Usage: "Determines the changes needed to update the Docker host to match the deployments " +
+				"specified by the local pallet",
 			Action: planAction,
 		},
 		{
 			Name:     "apply",
-			Category: "Use the environment",
-			Usage: "Updates the Docker Swarm to match the deployments specified by the " +
-				"local environment",
-			Action: applyAction,
+			Category: "Use the pallet",
+			Usage:    "Updates the Docker host to match the deployments specified by the local pallet",
+			Action:   applyAction,
 		},
 		{
 			Name:     "show",
-			Category: "Query the environment",
-			Usage:    "Describes the local environment",
+			Category: "Query the pallet",
+			Usage:    "Describes the local pallet",
 			Action:   showAction,
 		},
 		{
 			Name:     "ls-repo",
 			Aliases:  []string{"list-repositories"},
-			Category: "Query the environment",
-			Usage:    "Lists repos available in the local environment",
+			Category: "Query the pallet",
+			Usage:    "Lists repos available in the local pallet",
 			Action:   lsRepoAction,
 		},
 		{
 			Name:      "show-repo",
 			Aliases:   []string{"show-repository"},
-			Category:  "Query the environment",
-			Usage:     "Describes a repo available in the local environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a repo available in the local pallet",
 			ArgsUsage: "repo_path",
 			Action:    showRepoAction,
 		},
 		{
 			Name:     "ls-pkg",
 			Aliases:  []string{"list-packages"},
-			Category: "Query the environment",
-			Usage:    "Lists packages available in the local environment",
+			Category: "Query the pallet",
+			Usage:    "Lists packages available in the local pallet",
 			Action:   lsPkgAction,
 		},
 		{
 			Name:      "show-pkg",
 			Aliases:   []string{"show-package"},
-			Category:  "Query the environment",
-			Usage:     "Describes a package available in the local environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a package available in the local pallet",
 			ArgsUsage: "package_path",
 			Action:    showPkgAction,
 		},
 		{
 			Name:     "ls-depl",
 			Aliases:  []string{"list-deployments"},
-			Category: "Query the environment",
-			Usage:    "Lists package deployments specified by the local environment",
+			Category: "Query the pallet",
+			Usage:    "Lists package deployments specified by the local pallet",
 			Action:   lsDeplAction,
 		},
 		{
 			Name:      "show-depl",
 			Aliases:   []string{"show-deployment"},
-			Category:  "Query the environment",
-			Usage:     "Describes a package deployment specified by the local environment",
+			Category:  "Query the pallet",
+			Usage:     "Describes a package deployment specified by the local pallet",
 			ArgsUsage: "package_path",
 			Action:    showDeplAction,
 		},
 		// {
 		// 	Name:      "add-repo",
 		// 	Aliases:   []string{"add-repositories"},
-		// 	Category:  "Modify the environment",
-		// 	Usage:     "Adds repos to the environment, tracking specified versions or branches",
+		// 	Category:  "Modify the pallet",
+		// 	Usage:     "Adds repos to the pallet, tracking specified versions or branches",
 		// 	ArgsUsage: "[repo_path@version_query]...",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("adding repos", c.Args())
@@ -153,8 +152,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:      "rm-repo",
 		// 	Aliases:   []string{"remove-repositories"},
-		// 	Category:  "Modify the environment",
-		// 	Usage:     "Removes a repo from the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage:     "Removes a repo from the pallet",
 		// 	ArgsUsage: "repo_path",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("removing repo", c.Args().First())
@@ -164,8 +163,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:  "add-depl",
 		// 	Aliases:   []string{"add-deployments"},
-		// 	Category:  "Modify the environment",
-		// 	Usage: "Adds a package deployment to the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage: "Adds a package deployment to the pallet",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("adding package deployment", c.Args().First())
 		// 		return nil
@@ -174,8 +173,8 @@ var Cmd = &cli.Command{
 		// {
 		// 	Name:      "rm-depl",
 		// 	Aliases:   []string{"remove-deployments"},
-		// 	Category:  "Modify the environment",
-		// 	Usage:     "Removes a package deployment from the environment",
+		// 	Category:  "Modify the pallet",
+		// 	Usage:     "Removes a package deployment from the pallet",
 		// 	ArgsUsage: "package_path",
 		// 	Action: func(c *cli.Context) error {
 		// 		fmt.Println("removing package deployment", c.Args().First())
@@ -187,11 +186,11 @@ var Cmd = &cli.Command{
 
 // var remoteCmd = &cli.Command{
 // 	Name:  "remote",
-// 	Usage: "Manages the local environment's relationship to the remote source",
+// 	Usage: "Manages the local pallet's relationship to the remote source",
 // 	Subcommands: []*cli.Command{
 // 		{
 // 			Name:  "set",
-// 			Usage: "Sets the remote source for the local environment",
+// 			Usage: "Sets the remote source for the local pallet",
 // 			Action: func(c *cli.Context) error {
 // 				fmt.Println("setting remote source to", c.Args().First())
 // 				return nil