+++ cmd/forklift/plt/packages.go
@@ -1,4 +1,4 @@
-package env
+package plt
 
 import (
 	"github.com/urfave/cli/v2"
@@ -9,22 +9,22 @@ import (
 // ls-pkg
 
 func lsPkgAction(c *cli.Context) error {
-	env, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c, true)
 	if err != nil {
 		return err
 	}
 
-	return fcli.PrintEnvPkgs(0, env, cache)
+	return fcli.PrintPalletPkgs(0, pallet, cache)
 }
 
 // show-pkg
 
 func showPkgAction(c *cli.Context) error {
-	env, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c, true)
 	if err != nil {
 		return err
 	}
 
 	pkgPath := c.Args().First()
-	return fcli.PrintPkgInfo(0, env, cache, pkgPath)
+	return fcli.PrintPkgInfo(0, pallet, cache, pkgPath)
 }