+++ cmd/forklift/plt/pallets.go
@@ -1,7 +1,8 @@
-package env
+package plt
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
@@ -11,41 +12,39 @@ import (
 	"github.com/PlanktoScope/forklift/internal/clients/git"
 )
 
-var errMissingEnv = errors.Errorf(
-	"you first need to set up a local environment with `forklift env clone`",
-)
-
 func processFullBaseArgs(
 	c *cli.Context, ensureCache bool,
-) (env *forklift.FSEnv, cache *forklift.FSRepoCache, err error) {
+) (pallet *forklift.FSPallet, cache *forklift.FSRepoCache, err error) {
 	workspace, err := forklift.LoadWorkspace(c.String("workspace"))
 	if err != nil {
 		return nil, nil, err
 	}
-	if env, err = workspace.GetCurrentEnv(); err != nil {
-		return nil, nil, errMissingEnv
+	if pallet, err = getPallet(c.String("workspace")); err != nil {
+		return nil, nil, err
 	}
 	if cache, err = workspace.GetRepoCache(); err != nil {
 		return nil, nil, err
 	}
 	if ensureCache && !cache.Exists() {
 		return nil, nil, errors.New(
-			"you first need to cache the repos specified by your environment with " +
-				"`forklift env cache-repo`",
+			"you first need to cache the repos specified by your pallet with `forklift plt cache-repo`",
 		)
 	}
-	return env, cache, nil
+	return pallet, cache, nil
 }
 
-func getEnv(wpath string) (env *forklift.FSEnv, err error) {
+func getPallet(wpath string) (pallet *forklift.FSPallet, err error) {
 	workspace, err := forklift.LoadWorkspace(wpath)
 	if err != nil {
 		return nil, err
 	}
-	if env, err = workspace.GetCurrentEnv(); err != nil {
-		return nil, errMissingEnv
+	if pallet, err = workspace.GetCurrentPallet(); err != nil {
+		return nil, errors.Wrapf(
+			err, "couldn't load local pallet from workspace (you may need to first set up a local "+
+				"pallet with `forklift plt clone`)",
+		)
 	}
-	return env, nil
+	return pallet, nil
 }
 
 // clone
@@ -68,38 +67,38 @@ func cloneAction(c *cli.Context) error {
 	if err != nil {
 		return errors.Wrap(err, "couldn't load workspace")
 	}
-	local := workspace.GetCurrentEnvPath()
-	fmt.Printf("Cloning environment %s to %s...\n", remote, local)
+	local := workspace.GetCurrentPalletPath()
+	fmt.Printf("Cloning pallet %s to %s...\n", remote, local)
 	gitRepo, err := git.Clone(remote, local)
 	if err != nil {
 		if !errors.Is(err, git.ErrRepositoryAlreadyExists) {
 			return errors.Wrapf(
-				err, "couldn't clone environment %s at release %s to %s", remote, release, local,
+				err, "couldn't clone pallet %s at release %s to %s", remote, release, local,
 			)
 		}
 		if !c.Bool("force") {
 			return errors.Wrap(
 				err,
-				"you need to first delete your local environment with `forklift env rm` before "+
+				"you need to first delete your local pallet with `forklift plt rm` before "+
 					"cloning another remote release to it",
 			)
 		}
 
-		env, eerr := workspace.GetCurrentEnv()
-		if eerr != nil {
+		pallet, perr := workspace.GetCurrentPallet()
+		if perr != nil {
 			return err
 		}
 		fmt.Println(
-			"Removing local environment from workspace, because it already exists and the " +
+			"Removing local pallet from workspace, because it already exists and the " +
 				"command's --force flag was enabled...",
 		)
-		if err = env.Remove(); err != nil {
-			return errors.Wrap(err, "couldn't remove local environment")
+		if err = pallet.Remove(); err != nil {
+			return errors.Wrap(err, "couldn't remove local pallet")
 		}
-		fmt.Printf("Cloning environment %s to %s...\n", remote, env.FS.Path())
-		if gitRepo, err = git.Clone(remote, env.FS.Path()); err != nil {
+		fmt.Printf("Cloning pallet %s to %s...\n", remote, pallet.FS.Path())
+		if gitRepo, err = git.Clone(remote, pallet.FS.Path()); err != nil {
 			return errors.Wrapf(
-				err, "couldn't clone environment %s at release %s to %s", remote, release, env.FS.Path(),
+				err, "couldn't clone pallet %s at release %s to %s", remote, release, pallet.FS.Path(),
 			)
 		}
 	}
@@ -107,20 +106,21 @@ func cloneAction(c *cli.Context) error {
 	if err = gitRepo.Checkout(release); err != nil {
 		return errors.Wrapf(err, "couldn't check out release %s at %s", release, local)
 	}
-	fmt.Println("Done! Next, you'll probably want to run `forklift env cache-repo`.")
+	fmt.Println("Done! Next, you'll probably want to run `forklift plt cache-repo`.")
 	return nil
 }
 
 // fetch
 
 func fetchAction(c *cli.Context) error {
-	env, err := getEnv(c.String("workspace"))
+	workspace, err := forklift.LoadWorkspace(c.String("workspace"))
 	if err != nil {
 		return err
 	}
+	palletPath := workspace.GetCurrentPalletPath()
 
 	fmt.Println("Fetching updates...")
-	updated, err := git.Fetch(env.FS.Path())
+	updated, err := git.Fetch(palletPath)
 	if err != nil {
 		return errors.Wrap(err, "couldn't fetch changes from the remote release")
 	}
@@ -134,15 +134,16 @@ func fetchAction(c *cli.Context) error {
 // pull
 
 func pullAction(c *cli.Context) error {
-	env, err := getEnv(c.String("workspace"))
+	workspace, err := forklift.LoadWorkspace(c.String("workspace"))
 	if err != nil {
 		return err
 	}
+	palletPath := workspace.GetCurrentPalletPath()
 
-	fmt.Println("Attempting to fast-forward the local environment...")
-	updated, err := git.Pull(env.FS.Path())
+	fmt.Println("Attempting to fast-forward the local pallet...")
+	updated, err := git.Pull(palletPath)
 	if err != nil {
-		return errors.Wrap(err, "couldn't fast-forward the local environment")
+		return errors.Wrap(err, "couldn't fast-forward the local pallet")
 	}
 	if !updated {
 		fmt.Println("No changes from the remote release.")
@@ -154,36 +155,37 @@ func pullAction(c *cli.Context) error {
 // rm
 
 func rmAction(c *cli.Context) error {
-	env, err := getEnv(c.String("workspace"))
+	workspace, err := forklift.LoadWorkspace(c.String("workspace"))
 	if err != nil {
 		return err
 	}
+	palletPath := workspace.GetCurrentPalletPath()
 
-	fmt.Printf("Removing local environment from workspace...\n")
+	fmt.Printf("Removing local pallet from workspace...\n")
 	// TODO: return an error if there are uncommitted or unpushed changes to be removed - in which
 	// case require a --force flag
-	return errors.Wrap(env.Remove(), "couldn't remove local environment")
+	return errors.Wrap(os.RemoveAll(palletPath), "couldn't remove local pallet")
 }
 
 // show
 
 func showAction(c *cli.Context) error {
-	env, err := getEnv(c.String("workspace"))
+	pallet, err := getPallet(c.String("workspace"))
 	if err != nil {
 		return err
 	}
-	return fcli.PrintEnvInfo(0, env)
+	return fcli.PrintPalletInfo(0, pallet)
 }
 
 // check
 
 func checkAction(c *cli.Context) error {
-	env, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c, true)
 	if err != nil {
 		return err
 	}
 
-	if err := fcli.CheckEnv(0, env, cache); err != nil {
+	if err := fcli.CheckPallet(0, pallet, cache); err != nil {
 		return err
 	}
 	return nil
@@ -192,14 +194,14 @@ func checkAction(c *cli.Context) error {
 // plan
 
 func planAction(c *cli.Context) error {
-	env, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c, true)
 	if err != nil {
 		return err
 	}
 
-	if err := fcli.PlanEnv(0, env, cache); err != nil {
+	if err := fcli.PlanPallet(0, pallet, cache); err != nil {
 		return errors.Wrap(
-			err, "couldn't deploy local environment (have you run `forklift env cache` recently?)",
+			err, "couldn't deploy local pallet (have you run `forklift plt cache` recently?)",
 		)
 	}
 	return nil
@@ -208,14 +210,14 @@ func planAction(c *cli.Context) error {
 // apply
 
 func applyAction(c *cli.Context) error {
-	env, cache, err := processFullBaseArgs(c, true)
+	pallet, cache, err := processFullBaseArgs(c, true)
 	if err != nil {
 		return err
 	}
 
-	if err := fcli.ApplyEnv(0, env, cache); err != nil {
+	if err := fcli.ApplyPallet(0, pallet, cache); err != nil {
 		return errors.Wrap(
-			err, "couldn't deploy local environment (have you run `forklift env cache` recently?)",
+			err, "couldn't deploy local pallet (have you run `forklift plt cache` recently?)",
 		)
 	}
 	fmt.Println()